backend/requirements.txt
pandas==2.1.4
supabase==2.3.4
python-dotenv==1.0.0
requests==2.31.0

